Improved Board infrastructure
Add magic methods to handle indexing of Board instances
using a position tuple, used either to read or write to
the internal state.

Added/cleaned up some validation code

Differentiated between gamestate and boardstate. The latter
being the nested list representing the board. While the former
being a tuple containing the latter, in addition to a value
indicating which player is to move.